Remove extraneous comments and refactor managers into dedicated folder
