+++ Managers/TokenManager.js
