+++ Managers/UserManager.js
