+++ User-Authentication/SignUp.js
@@ -7,9 +7,9 @@ import { useNavigation } from '@react-navigation/native';
 import { useState } from 'react';
 import { GetJWTTokenRequest, AddUserRequest, GetUserByIDRequest, UpdateUserInfoRequest } from "../gen/proto/users_pb";
 import { UsersClient } from "../gen/proto/UsersServiceClientPb";
-import TokenManager from './TokenManager';
-import UserManager from './UserManager';
-import UsersClientManager from './UsersClientManager.js';
+import TokenManager from "../Managers/TokenManager";
+import UsersClientManager from "../Managers/UsersClientManager";
+import UserManager from '../Managers/UserManager';
 import { Date } from "../gen/proto/common_pb";
 import KIC_Style from "../Components/Style";
 import { Text, TouchableOpacity, Image, View, TextInput } from "react-native";
